chore: modularize main.py into app/ package
First steps of app modularization (the 'all code in main.py' rule is
retired). main.py becomes a thin uvicorn entrypoint + compatibility
shim; the application now lives in layered modules:

  app/config.py        paths, env, runtime AI config (replaces app.state)
  app/sidecars.py      schema, path safety, sidecar/metadata I/O
  app/ai_metadata.py   OpenAI prompt/request/populate pipeline
  app/watcher.py       pending-file detection and background poll
  app/security.py      basic auth dependency + security headers
  app/routes_admin.py  all /admin routes (APIRouter)
  app/routes_public.py gallery + artwork detail (APIRouter)
  app/factory.py       create_app: lifespan, mounts, middleware, routers

Cross-module calls go through module attributes so tests patch one
location; the dead duplicate FastAPI instance is removed. Test
monkeypatch targets updated to the defining modules; adds a shim
compat test and a route-table regression test.
